Improve tests and utils around public schema (#118)
I added the link utils Jeff created (`extractLinkImplementations`).
I changed the logic of `transformSupergraphToPublicSchema` and
`containsSupergraphSpec` to reuse what is defined in
`src/specifications`. The supergraph to public schema transformation
supports now aliasing.

I added quite a lot of tests to check if the `publicSdl` we're producing
is correct.
The `assertCompositionSuccess` function that is used in tests, builds a
GraphQLSchema instance from the `publicSdl` (in case of Apollo it uses
their own api to achieve the same thing) and validates it. This way we
can assume the `publicSdl` is valid and executable.